requirements.txt
streamlit
sqlalchemy
pyodbc
pandas
matplotlib
plotly.express